Extract consistency settings into a separate class.
The idea is to refer to it from different configuration case classes, i.e. from
`EventualCassandraConfig`, `SnapshotCassandraConfig` and
`CassandraHealthCheck`.
