+++ eventual-cassandra/src/main/scala/com/evolutiongaming/kafka/journal/eventual/cassandra/EventualCassandraConfig.scala
@@ -40,7 +40,7 @@ final case class EventualCassandraConfig(
       fetchSize = 1000,
       defaultIdempotence = true)),
   schema: SchemaConfig = SchemaConfig.default,
-  consistencyConfig: EventualCassandraConfig.ConsistencyConfig = EventualCassandraConfig.ConsistencyConfig.default)
+  consistencyConfig: ConsistencyConfig = ConsistencyConfig.default)
 
 object EventualCassandraConfig {
 
@@ -48,33 +48,4 @@ object EventualCassandraConfig {
 
   implicit val configReaderEventualCassandraConfig: ConfigReader[EventualCassandraConfig] = deriveReader
 
-
-  final case class ConsistencyConfig(
-    read: ConsistencyConfig.Read = ConsistencyConfig.Read.default,
-    write: ConsistencyConfig.Write = ConsistencyConfig.Write.default)
-
-  object ConsistencyConfig {
-
-    implicit val configReaderConsistencyConfig: ConfigReader[ConsistencyConfig] = deriveReader
-
-    val default: ConsistencyConfig = ConsistencyConfig()
-
-
-    final case class Read(value: ConsistencyLevel = ConsistencyLevel.LOCAL_QUORUM)
-
-    object Read {
-      val default: Read = Read()
-
-      implicit val configReaderRead: ConfigReader[Read] = ConfigReader[ConsistencyLevel].map { a => Read(a) }
-    }
-
-
-    final case class Write(value: ConsistencyLevel = ConsistencyLevel.LOCAL_QUORUM)
-
-    object Write {
-      val default: Write = Write()
-
-      implicit val configReaderWrite: ConfigReader[Write] = ConfigReader[ConsistencyLevel].map { a => Write(a) }
-    }
-  }
 }
